Refactor: make generator independent from stringSearch
